+++ src/app/mod.rs
@@ -1145,6 +1145,7 @@ impl App {
                         cwd: None,
                         focus: true,
                         label,
+                        index: None,
                         env: Default::default(),
                     },
                 );