+++ src/server/headless.rs
@@ -1058,6 +1058,7 @@ impl HeadlessServer {
                 cwd: None,
                 focus: true,
                 label,
+                index: None,
                 env: Default::default(),
             }),
         )